+++ Nickel/Framework/Patches/ArtifactRewardPatches.cs
@@ -56,7 +56,7 @@ internal static class ArtifactRewardPatches
 		}
 		catch (Exception ex)
 		{
-			Nickel.Instance.ModManager.Logger.LogCritical("Could not patch method {Method} - {ModLoaderName} probably won't work.\nReason: {Exception}", originalMethod, NickelConstants.Name, ex);
+			Nickel.Instance.ModManager.Logger.LogCritical("Could not patch method {DeclaringType}::{Method} - {ModLoaderName} probably won't work.\nReason: {Exception}", originalMethod.DeclaringType, originalMethod, NickelConstants.Name, ex);
 			return instructions;
 		}
 	}